% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Figure1.R
\name{Figure1a_Plot_ClustersofClusters_CentroidClustering}
\alias{Figure1a_Plot_ClustersofClusters_CentroidClustering}
\title{Plot Figure 1a Middle and Bottom Panel, clusters of clusters and centroid clustering plot # plot only}
\usage{
Figure1a_Plot_ClustersofClusters_CentroidClustering()
}
\description{
Plot Figure 1a Middle and Bottom Panel, clusters of clusters and centroid clustering plot # plot only
}
