% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Figure1.R
\name{Figure1b_Classification_CpGpredictor}
\alias{Figure1b_Classification_CpGpredictor}
\title{Plot Figure 1b right Panel, classification of retinoblastoma by omics/9-CpG predictor}
\usage{
Figure1b_Classification_CpGpredictor()
}
\description{
Plot Figure 1b right Panel, classification of retinoblastoma by omics/9-CpG predictor
}
