% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Figure1.R
\name{Figure1b_Heatmap_9CpG_metharray}
\alias{Figure1b_Heatmap_9CpG_metharray}
\title{Plot Figure 1b left Panel, heatmap of 9 CpGs # to be finished}
\usage{
Figure1b_Heatmap_9CpG_metharray()
}
\description{
Plot Figure 1b left Panel, heatmap of 9 CpGs # to be finished
}
