% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Figure1.R
\name{Figure1b_Scatter_CpGarray_pyroseq}
\alias{Figure1b_Scatter_CpGarray_pyroseq}
\title{Plot Figure 1b middle Panel, scatter of 9 CpGs correlating CpGarray and pyroseq}
\usage{
Figure1b_Scatter_CpGarray_pyroseq()
}
\description{
Plot Figure 1b middle Panel, scatter of 9 CpGs correlating CpGarray and pyroseq
}
