% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Figure1.R
\name{Figure1c_Plot_Clinical}
\alias{Figure1c_Plot_Clinical}
\title{Plot Figure 1c, Heatmap of different clinical features in two subtypes of retinoblastoma}
\usage{
Figure1c_Plot_Clinical(input)
}
\description{
Plot Figure 1c, Heatmap of different clinical features in two subtypes of retinoblastoma
}
