% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Figure2.R
\name{Figure2efg_Meth_FigureS2_Meth}
\alias{Figure2efg_Meth_FigureS2_Meth}
\title{Plot Figure 2e heatmap of differential methylated CpGs}
\usage{
Figure2efg_Meth_FigureS2_Meth()
}
\description{
Plot Figure 2e heatmap of differential methylated CpGs
}
