% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Figure3.R
\name{Figure3a_volcanoPlot}
\alias{Figure3a_volcanoPlot}
\title{Plot Figure 3a volcano plot of differential expressed genes between 2 subtypes}
\usage{
Figure3a_volcanoPlot(expRBRET)
}
\description{
Plot Figure 3a volcano plot of differential expressed genes between 2 subtypes
}
